user/requirements.txt
pymongo==3.9.0
Flask_Cors==3.0.8
Flask==1.0.2
PyJWT==1.7.1
python-dotenv==0.12.0
python_bcrypt==0.3.2
dnspython